Module type Tyabt.Operator

An operator is a function symbol.

type ('arity, 'sort) t

The operator's type contains two phantom type parameters, the first for the operator's arity and the second for the operator's sort.

val equal : ('arity1, 'sort) t -> ('arity2, 'sort) t -> ('arity1, 'arity2) eq option

Checks the equality of two operators from the same sort. Iff the operators are equal, returns a proof that their arities are equal.

val pp_print : Stdlib.Format.formatter -> (_, _) t -> unit

Prettyprints the operator.